+++ pythainlp/benchmarks/__init__.py
@@ -3,10 +3,17 @@
 # SPDX-License-Identifier: Apache-2.0
 """Performance benchmarking."""
 
-__all__: list[str] = ["benchmark", "bleu_score", "rouge_score", "word_error_rate"]
+__all__: list[str] = [
+    "benchmark",
+    "bleu_score",
+    "character_error_rate",
+    "rouge_score",
+    "word_error_rate",
+]
 
 from pythainlp.benchmarks.metrics import (
     bleu_score,
+    character_error_rate,
     rouge_score,
     word_error_rate,
 )